+++ tntsearch.php
@@ -110,7 +110,7 @@ class TNTSearchPlugin extends Plugin
             $scheduler = $e['scheduler'];
             $at = $this->config->get('plugins.tntsearch.scheduled_index.at');
             $logs = $this->config->get('plugins.tntsearch.scheduled_index.logs');
-            $job = $scheduler->addFunction('Grav\Plugin\TNTSearchPlugin::indexJob', [], 'tntsearch-index');
+            $job = $scheduler->addCommand('bin/plugin tntsearch index', [], 'tntsearch-index');
             $job->at($at);
             $job->output($logs);
             $job->backlink('/plugins/tntsearch');
@@ -434,7 +434,11 @@ class TNTSearchPlugin extends Plugin
         throw new \RuntimeException('Search class: ' . $type . ' does not exist');
     }
 
-    public static function indexJob()
+    /**
+     * @param string|null $langCode
+     * @return array
+     */
+    public static function indexJob(string $langCode = null)
     {
         $grav = Grav::instance();
         $grav['debugger']->enabled(false);
@@ -445,15 +449,34 @@ class TNTSearchPlugin extends Plugin
             $pages->enablePages();
         }
 
-        /** @var Language $language */
-        $language = $grav['language'];
-
         ob_start();
 
-        if ($language->enabled()) {
-            foreach ($language->getLanguages() as $lang) {
-                $language->init();
-                $language->setActive($lang);
+        /** @var Language $language */
+        $language = $grav['language'];
+        $langEnabled = $language->enabled();
+
+        // TODO: can be removed when Grav minimum >= v1.6.22
+        $hasReset = method_exists($pages, 'reset');
+        if (!$hasReset && !$langCode) {
+            $langCode = $language->getActive();
+        }
+
+        if ($langCode && (!$langEnabled || !$language->validate($langCode))) {
+            $langCode = null;
+        }
+
+        $langCodes = $langCode ? [$langCode] : $language->getLanguages();
+        if ($langCodes) {
+            foreach ($langCodes as $lang) {
+                if ($lang !== $language->getActive()) {
+                    $language->init();
+                    $language->setActive($lang);
+
+                    // TODO: $hasReset test can be removed (keep reset!) when Grav minimum >= v1.6.22
+                    if ($hasReset) {
+                        $pages->reset();
+                    }
+                }
 
                 echo "\nLanguage: {$lang}\n";