+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# v3.2.0
+## 06/08/2020
+
+1. [](#new)
+    * Added support for CLI `bin/plugin index` to index only a single language (`--language=en`)
+1. [](#improved)
+    * Renamed CLI classes to avoid class name conflicts
+1. [](#bugfix)
+    * Fixed non-routable and non-published pages showing up in search results
+    * Fixed indexing in multi-language sites
+    * Use CLI command directly in scheduler command to work [#95](https://github.com/trilbymedia/grav-plugin-tntsearch/issues/95) 
+
 # v3.1.1
 ## 02/12/2020